+++ lib/rspec/core/option_parser.rb
@@ -5,11 +5,18 @@ module RSpec::Core
   # @private
   class Parser
     def self.parse(args)
-      new.parse(args)
+      new(args).parse
     end
 
-    def parse(args)
-      return {} if args.empty?
+    attr_reader :original_args
+
+    def initialize(original_args)
+      @original_args = original_args
+    end
+
+    def parse
+      return { :files_or_directories_to_run => [] } if original_args.empty?
+      args = original_args.dup
 
       options = args.delete('--tty') ? { :tty => true } : {}
       begin
@@ -18,8 +25,11 @@ module RSpec::Core
         abort "#{e.message}\n\nPlease use --help for a listing of valid options"
       end
 
+      options[:files_or_directories_to_run] = args
       options
     end
+
+  private
 
     # rubocop:disable MethodLength
     def parser(options)
@@ -52,12 +62,13 @@ module RSpec::Core
           options[:order] = "rand:#{seed}"
         end
 
-        parser.on('--fail-fast', 'Abort the run on first failure.') do |_o|
-          options[:fail_fast] = true
+        parser.on('--bisect [verbose]', 'Repeatedly runs the suite in order to isolate the failures to the ',
+                  '  smallest reproducible case.') do |argument|
+          bisect_and_exit(argument)
         end
 
-        parser.on('--no-fail-fast', 'Do not abort the run on first failure.') do |_o|
-          options[:fail_fast] = false
+        parser.on('--[no-]fail-fast', 'Abort the run on first failure.') do |value|
+          set_fail_fast(options, value)
         end
 
         parser.on('--failure-exit-code CODE', Integer,
@@ -79,9 +90,7 @@ module RSpec::Core
         end
 
         parser.on('--init', 'Initialize your project with RSpec.') do |_cmd|
-          RSpec::Support.require_rspec_core "project_initializer"
-          ProjectInitializer.new.run
-          exit
+          initialize_project_and_exit
         end
 
         parser.separator("\n  **** Output ****\n\n")
@@ -156,6 +165,17 @@ module RSpec::Core
 
 FILTERING
 
+        parser.on('--only-failures', "Filter to just the examples that failed the last time they ran.") do
+          configure_only_failures(options)
+        end
+
+        parser.on("--next-failure", "Apply `--only-failures` and abort after one failure.",
+                  "  (Equivalent to `--only-failures --fail-fast --order defined`)") do
+          configure_only_failures(options)
+          set_fail_fast(options, true)
+          options[:order] ||= 'defined'
+        end
+
         parser.on('-P', '--pattern PATTERN', 'Load files matching pattern (default: "spec/**/*_spec.rb").') do |o|
           options[:pattern] = o
         end
@@ -180,18 +200,19 @@ FILTERING
           name, value = tag.gsub(/^(~@|~|@)/, '').split(':', 2)
           name = name.to_sym
 
-          options[filter_type] ||= {}
-          options[filter_type][name] = case value
-                                       when  nil        then true # The default value for tags is true
-                                       when 'true'      then true
-                                       when 'false'     then false
-                                       when 'nil'       then nil
-                                       when /^:/        then value[1..-1].to_sym
-                                       when /^\d+$/     then Integer(value)
-                                       when /^\d+.\d+$/ then Float(value)
-                                       else
-                                         value
-                                       end
+          parsed_value = case value
+                         when  nil        then true # The default value for tags is true
+                         when 'true'      then true
+                         when 'false'     then false
+                         when 'nil'       then nil
+                         when /^:/        then value[1..-1].to_sym
+                         when /^\d+$/     then Integer(value)
+                         when /^\d+.\d+$/ then Float(value)
+                         else
+                           value
+                         end
+
+          add_tag_filter(options, filter_type, name, parsed_value)
         end
 
         parser.on('--default-path PATH', 'Set the default path where RSpec looks for examples (can',
@@ -202,8 +223,7 @@ FILTERING
         parser.separator("\n  **** Utility ****\n\n")
 
         parser.on('-v', '--version', 'Display the version.') do
-          puts RSpec::Core::Version::STRING
-          exit
+          print_version_and_exit
         end
 
         # These options would otherwise be confusing to users, so we forcibly
@@ -215,9 +235,7 @@ FILTERING
         invalid_options = %w[-d --I]
 
         parser.on_tail('-h', '--help', "You're looking at it.") do
-          # Removing the blank invalid options from the output.
-          puts parser.to_s.gsub(/^\s+(#{invalid_options.join('|')})\s*$\n/, '')
-          exit
+          print_help_and_exit(parser, invalid_options)
         end
 
         # This prevents usage of the invalid_options.
@@ -230,5 +248,52 @@ FILTERING
       end
     end
     # rubocop:enable MethodLength
+
+    def add_tag_filter(options, filter_type, tag_name, value=true)
+      (options[filter_type] ||= {})[tag_name] = value
+    end
+
+    def set_fail_fast(options, value)
+      options[:fail_fast] = value
+    end
+
+    def configure_only_failures(options)
+      options[:only_failures] = true
+      add_tag_filter(options, :inclusion_filter, :last_run_status, 'failed')
+    end
+
+    def initialize_project_and_exit
+      RSpec::Support.require_rspec_core "project_initializer"
+      ProjectInitializer.new.run
+      exit
+    end
+
+    def bisect_and_exit(argument)
+      RSpec::Support.require_rspec_core "bisect/coordinator"
+
+      success = Bisect::Coordinator.bisect_with(
+        original_args,
+        RSpec.configuration,
+        bisect_formatter_for(argument)
+      )
+
+      exit(success ? 0 : 1)
+    end
+
+    def bisect_formatter_for(argument)
+      return Formatters::BisectDebugFormatter if argument == "verbose"
+      Formatters::BisectProgressFormatter
+    end
+
+    def print_version_and_exit
+      puts RSpec::Core::Version::STRING
+      exit
+    end
+
+    def print_help_and_exit(parser, invalid_options)
+      # Removing the blank invalid options from the output.
+      puts parser.to_s.gsub(/^\s+(#{invalid_options.join('|')})\s*$\n/, '')
+      exit
+    end
   end
 end