+++ lib/rspec/core/world.rb
@@ -15,19 +15,8 @@ module RSpec
         @example_groups = []
         @example_group_counts_by_spec_file = Hash.new(0)
         @filtered_examples = Hash.new do |hash, group|
-          hash[group] = begin
-            examples = group.examples.dup
-            examples = filter_manager.prune(examples)
-            examples.uniq!
-            examples
-          end
+          hash[group] = filter_manager.prune(group.examples)
         end
-      end
-
-      # @private
-      # Used internally to clear remaining groups when fail_fast is set.
-      def clear_remaining_example_groups
-        example_groups.clear
       end
 
       # @api private
@@ -112,6 +101,7 @@ module RSpec
       #
       # Notify reporter of filters.
       def announce_filters
+        fail_if_config_and_cli_options_invalid
         filter_announcements = []
 
         announce_inclusion_filter filter_announcements
@@ -136,13 +126,7 @@ module RSpec
         example_groups.clear
         if filter_manager.empty?
           reporter.message("No examples found.")
-        elsif exclusion_filter.empty?
-          message = everything_filtered_message
-          if @configuration.run_all_when_everything_filtered?
-            message << "; ignoring #{inclusion_filter.description}"
-          end
-          reporter.message(message)
-        elsif inclusion_filter.empty?
+        elsif exclusion_filter.empty? || inclusion_filter.empty?
           reporter.message(everything_filtered_message)
         end
       end
@@ -174,6 +158,16 @@ module RSpec
 
       def declaration_line_numbers
         @declaration_line_numbers ||= FlatMap.flat_map(example_groups, &:declaration_line_numbers)
+      end
+
+      def fail_if_config_and_cli_options_invalid
+        return unless @configuration.only_failures_but_not_configured?
+
+        reporter.abort_with(
+          "\nTo use `--only-failures`, you must first set " \
+          "`config.example_status_persistence_file_path`.",
+          1 # exit code
+        )
       end
     end
   end