+++ spec/rspec/core/formatters/json_formatter_spec.rb
@@ -13,7 +13,13 @@ require 'rspec/core/reporter'
 RSpec.describe RSpec::Core::Formatters::JsonFormatter do
   include FormatterSupport
 
-  it "outputs json (brittle high level functional test)" do
+  it "can be loaded via `--format json`" do
+    output = run_example_specs_with_formatter("json", false)
+    parsed = JSON.parse(output)
+    expect(parsed.keys).to include("examples", "summary", "summary_line")
+  end
+
+  it "outputs expected json (brittle high level functional test)" do
     group = RSpec.describe("one apiece") do
       it("succeeds") { expect(1).to eq 1 }
       it("fails") { fail "eek" }
@@ -35,6 +41,7 @@ RSpec.describe RSpec::Core::Formatters::JsonFormatter do
     this_file = relative_path(__FILE__)
 
     expected = {
+      :version => RSpec::Core::Version::STRING,
       :examples => [
         {
           :description => "succeeds",
@@ -75,7 +82,7 @@ RSpec.describe RSpec::Core::Formatters::JsonFormatter do
       :summary_line => "3 examples, 1 failure, 1 pending"
     }
     expect(formatter.output_hash).to eq expected
-    expect(output.string).to eq expected.to_json
+    expect(formatter_output.string).to eq expected.to_json
   end
 
   describe "#stop" do
@@ -87,9 +94,11 @@ RSpec.describe RSpec::Core::Formatters::JsonFormatter do
 
   describe "#close" do
     it "outputs the results as a JSON string" do
-      expect(output.string).to eq ""
+      expect(formatter_output.string).to eq ""
       send_notification :close, null_notification
-      expect(output.string).to eq("{}")
+      expect(formatter_output.string).to eq({
+        :version => RSpec::Core::Version::STRING
+      }.to_json)
     end
   end